+++ sharding-spring/sharding-jdbc-orchestration-spring/sharding-jdbc-orchestration-spring-namespace/src/test/java/org/apache/shardingsphere/shardingjdbc/orchestration/spring/registry/TestRegistryCenterRepositoryRepository.java
@@ -24,8 +24,8 @@ import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.orchestration.center.ConfigCenterRepository;
 import org.apache.shardingsphere.orchestration.center.RegistryCenterRepository;
-import org.apache.shardingsphere.orchestration.center.configuration.InstanceConfiguration;
 import org.apache.shardingsphere.orchestration.center.listener.DataChangedEventListener;
+import org.apache.shardingsphere.underlying.common.config.orchestration.CenterConfiguration;
 
 public final class TestRegistryCenterRepositoryRepository implements RegistryCenterRepository, ConfigCenterRepository {
     
@@ -34,7 +34,7 @@ public final class TestRegistryCenterRepositoryRepository implements RegistryCen
     private Properties properties;
     
     @Override
-    public void init(final InstanceConfiguration config) {
+    public void init(final CenterConfiguration config) {
     }
     
     @Override
